+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for CKEditor for Craft CMS
 
+## Unreleased
+
+- Added the “Parse embeds” field setting. ([#409](https://github.com/craftcms/ckeditor/pull/409))
+
 ## 4.8.0 - 2025-04-30
 
 - Added the “Field Limit” field setting, which can be set to a character limit or word limit. ([#384](https://github.com/craftcms/ckeditor/pull/384))